requirements.txt
pipecat-ai[speechmatics,elevenlabs,webrtc,qwen,moondream]>=0.0.48
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
loguru>=0.7.0
certifi>=2024.0.0

